requirements.txt
streamlit
pandas
numpy
requests
plotly
scikit-learn
python-dotenv
joblib
